fix(blocks-engine,blocks-react): publish the provenance type, and fix a false import match
The type is reachable now from every entry point that publishes the field
naming it — the engine root, the format subpath, and the renderer. A package
exporting `BlockNode` but not the union one of its fields holds leaves a
consumer able to read the value and unable to write its type. The renderer
already had a guard for exactly that and it caught the omission.

The predicate moves beside the type it describes rather than living inside the
op vocabulary, because a document reaches storage by more than one road and a
record one road admits and another refuses is one that exists and cannot be
edited. Wiring the second road is a separate change; see the PR discussion.

The format entry's boundary test scanned for imports with a pattern that a
string literal could satisfy: `"from"` ends in a quote, so the keyword matched
the word inside the string and the closing quote read as a specifier's opening
one, capturing the next two lines. A module specifier cannot contain a newline,
so excluding one refuses that without hiding any real import. Two tests: the
false positive, and a control that every real specifier shape is still seen.
