+++ packages/blocks-engine/src/index.ts
@@ -68,6 +68,7 @@ export type {
   Condition,
   DocumentFormatVersion,
   DocumentKind,
+  BlockOrigin,
   DocumentSettings,
   LocaleOverlay,
   LocaleOverlayValue,
