+++ packages/blocks-react/src/index.ts
@@ -247,6 +247,7 @@ export type {
   BlockIsland,
   BlockMigrationInfo,
   BlockNode,
+  BlockOrigin,
   BlockPart,
   BlockRenderResult,
   BlockSeoContribution,
